2to3 long: Rename long to int
    2to3 -w -n -f long ./src
    2to3 -w -n -f long ./tests

+++ tests/sentry/similarity/test_encoder.py
@@ -28,7 +28,7 @@ def test_builtin_types():
     ]
 
     try:
-        values.append(long(1))  # noqa
+        values.append(1)  # noqa
     except NameError:
         pass
